requirements.txt
# Web 界面
streamlit==1.31.0

# AI SDK
anthropic==0.18.1
google-generativeai==0.3.2

# Notion API
notion-client==2.2.1

# 数据处理
pandas==2.1.4
python-dotenv==1.0.0

# 定时任务
schedule==1.2.0

# HTTP 请求
requests==2.31.0
httpx==0.26.0

# RSS 解析
feedparser==6.0.11

# HTML 处理
beautifulsoup4==4.12.3
lxml==5.1.0

# Markdown 转 HTML
markdown==3.5.2
